+++ src/ts/Managers/OldCampaignRefreshManager.ts
@@ -7,29 +7,39 @@ import { PlacementState } from 'Models/Placement';
 import { Configuration } from 'Models/Configuration';
 import { Diagnostics } from 'Utilities/Diagnostics';
 import { AbstractAdUnit } from 'AdUnits/AbstractAdUnit';
-import { INativeResponse } from 'Utilities/Request';
+import { Request, INativeResponse } from 'Utilities/Request';
 import { Session } from 'Models/Session';
 import { FocusManager } from 'Managers/FocusManager';
 import { SdkStats } from 'Utilities/SdkStats';
 import { Platform } from 'Constants/Platform';
+import { RefreshManager } from 'Managers/RefreshManager';
+import { SessionManager } from 'Managers/SessionManager';
+import { JsonParser } from 'Utilities/JsonParser';
+import { StorageType } from 'Native/Api/Storage';
+import { ClientInfo } from 'Models/ClientInfo';
+import { Cache } from 'Utilities/Cache';
+import { OperativeEventManager } from 'Managers/OperativeEventManager';
 
-export class CampaignRefreshManager {
-    public static NoFillDelay = 3600;
-    public static ErrorRefillDelay = 3600;
-    public static ParsingErrorRefillDelay = 60;
-
+export class OldCampaignRefreshManager extends RefreshManager {
     private _nativeBridge: NativeBridge;
     private _wakeUpManager: WakeUpManager;
     private _campaignManager: CampaignManager;
     private _configuration: Configuration;
     private _currentAdUnit: AbstractAdUnit;
     private _focusManager: FocusManager;
+    private _sessionManager: SessionManager;
+    private _clientInfo: ClientInfo;
+    private _request: Request;
+    private _cache: Cache;
+    private _operativeEventManager: OperativeEventManager;
     private _refillTimestamp: number;
     private _needsRefill = true;
     private _refreshAllowed = true;
     private _campaignCount: number;
     private _parsingErrorCount: number;
     private _noFills: number;
+    private _mustReinitialize: boolean = false;
+    private _configJsonCheckedAt: number;
 
     // constant value that determines the delay for refreshing ads after backend has processed a start event
     // set to five seconds because backend should usually process start event in less than one second but
@@ -37,22 +47,32 @@ export class CampaignRefreshManager {
     // this constant is intentionally named "magic" constant because the value is only a best guess and not a real technical constant
     private _startRefreshMagicConstant: number = 5000;
 
-    constructor(nativeBridge: NativeBridge, wakeUpManager: WakeUpManager, campaignManager: CampaignManager, configuration: Configuration, focusManager: FocusManager) {
+    constructor(nativeBridge: NativeBridge, wakeUpManager: WakeUpManager, campaignManager: CampaignManager, configuration: Configuration, focusManager: FocusManager, sessionManager: SessionManager, clientInfo: ClientInfo, request: Request, cache: Cache, operativeEventManager: OperativeEventManager) {
+        super();
+
         this._nativeBridge = nativeBridge;
         this._wakeUpManager = wakeUpManager;
         this._campaignManager = campaignManager;
         this._configuration = configuration;
         this._focusManager = focusManager;
+        this._sessionManager = sessionManager;
+        this._clientInfo = clientInfo;
+        this._request = request;
+        this._cache = cache;
+        this._operativeEventManager = operativeEventManager;
         this._refillTimestamp = 0;
         this._campaignCount = 0;
         this._parsingErrorCount = 0;
         this._noFills = 0;
+        this._configJsonCheckedAt = Date.now();
+        this._mustReinitialize = false;
 
         this._campaignManager.onCampaign.subscribe((placementId, campaign) => this.onCampaign(placementId, campaign));
         this._campaignManager.onNoFill.subscribe(placementId => this.onNoFill(placementId));
         this._campaignManager.onError.subscribe((error, placementIds, session) => this.onError(error, placementIds, session));
         this._campaignManager.onConnectivityError.subscribe((placementIds) => this.onConnectivityError(placementIds));
         this._campaignManager.onAdPlanReceived.subscribe((refreshDelay, campaignCount) => this.onAdPlanReceived(refreshDelay, campaignCount));
+        this._wakeUpManager.onNetworkConnected.subscribe(() => this.onNetworkConnected());
         if(this._nativeBridge.getPlatform() === Platform.IOS) {
             this._focusManager.onAppForeground.subscribe(() => this.onAppForeground());
         } else {
@@ -71,6 +91,11 @@ export class CampaignRefreshManager {
     }
 
     public setCurrentAdUnit(adUnit: AbstractAdUnit): void {
+        this.shouldReinitialize().then((reinitialize) => {
+            this._mustReinitialize = reinitialize;
+            this.setRefreshAllowed(!reinitialize);
+        });
+
         this._currentAdUnit = adUnit;
         const onStartObserver = this._currentAdUnit.onStart.subscribe(() => {
             this._currentAdUnit.onStart.unsubscribe(onStartObserver);
@@ -194,10 +219,10 @@ export class CampaignRefreshManager {
         }, session);
         this._nativeBridge.Sdk.logError(JSON.stringify(error));
 
-        const minimumRefreshTimestamp = Date.now() + CampaignRefreshManager.ErrorRefillDelay * 1000;
+        const minimumRefreshTimestamp = Date.now() + RefreshManager.ErrorRefillDelay * 1000;
         if(this._refillTimestamp === 0 || this._refillTimestamp > minimumRefreshTimestamp) {
             this._refillTimestamp = minimumRefreshTimestamp;
-            this._nativeBridge.Sdk.logDebug('Unity Ads will refresh ads in ' + CampaignRefreshManager.ErrorRefillDelay + ' seconds');
+            this._nativeBridge.Sdk.logDebug('Unity Ads will refresh ads in ' + RefreshManager.ErrorRefillDelay + ' seconds');
         }
 
         if(this._currentAdUnit && this._currentAdUnit.isShowing()) {
@@ -214,10 +239,10 @@ export class CampaignRefreshManager {
         if(this._campaignCount === 1) {
             this._parsingErrorCount++;
 
-            if(this._parsingErrorCount === 1 && CampaignRefreshManager.ParsingErrorRefillDelay > 0) {
-                const retryDelaySeconds: number = CampaignRefreshManager.ParsingErrorRefillDelay + Math.random() * CampaignRefreshManager.ParsingErrorRefillDelay;
+            if(this._parsingErrorCount === 1 && RefreshManager.ParsingErrorRefillDelay > 0) {
+                const retryDelaySeconds: number = RefreshManager.ParsingErrorRefillDelay + Math.random() * RefreshManager.ParsingErrorRefillDelay;
                 this._nativeBridge.Sdk.logDebug('Unity Ads retrying failed campaign in ' + retryDelaySeconds + ' seconds');
-                this._refillTimestamp = Date.now() + CampaignRefreshManager.ParsingErrorRefillDelay * 1000;
+                this._refillTimestamp = Date.now() + RefreshManager.ParsingErrorRefillDelay * 1000;
                 setTimeout(() => {
                     this._nativeBridge.Sdk.logDebug('Unity Ads retrying failed campaign now');
                     this.refresh();
@@ -329,7 +354,13 @@ export class CampaignRefreshManager {
     }
 
     private onAdUnitClose(): void {
-        this.refresh();
+        this._nativeBridge.Sdk.logInfo('Closing Unity Ads ad unit');
+        if(this._mustReinitialize) {
+            this._nativeBridge.Sdk.logDebug('Unity Ads webapp has been updated, reinitializing Unity Ads');
+            this.reinitialize();
+        } else {
+            this.refresh();
+        }
     }
 
     private onAdUnitStartProcessed(): void {
@@ -340,5 +371,66 @@ export class CampaignRefreshManager {
                 }
             }, this._startRefreshMagicConstant);
         }
+    }
+
+    /*
+     Reinit logic copied from WebView.ts
+     */
+
+    private reinitialize() {
+        // save caching pause state in case of reinit
+        if(this._cache.isPaused()) {
+            Promise.all([this._nativeBridge.Storage.set(StorageType.PUBLIC, 'caching.pause.value', true), this._nativeBridge.Storage.write(StorageType.PUBLIC)]).then(() => {
+                this._nativeBridge.Sdk.reinitialize();
+            }).catch(() => {
+                this._nativeBridge.Sdk.reinitialize();
+            });
+        } else {
+            this._nativeBridge.Sdk.reinitialize();
+        }
+    }
+
+    private getConfigJson(): Promise<INativeResponse> {
+        return this._request.get(this._clientInfo.getConfigUrl() + '?ts=' + Date.now() + '&sdkVersion=' + this._clientInfo.getSdkVersion());
+    }
+
+    private shouldReinitialize(): Promise<boolean> {
+        if(!this._clientInfo.getWebviewHash()) {
+            return Promise.resolve(false);
+        }
+        if(Date.now() - this._configJsonCheckedAt <= 15 * 60 * 1000) {
+            return Promise.resolve(false);
+        }
+        return this.getConfigJson().then(response => {
+            this._configJsonCheckedAt = Date.now();
+            const configJson = JsonParser.parse(response.response);
+            return configJson.hash !== this._clientInfo.getWebviewHash();
+        }).catch((error) => {
+            return false;
+        });
+    }
+
+    /*
+     Connectivity handlers copied from WebView.ts
+     */
+
+    private onNetworkConnected() {
+        if(this._currentAdUnit && this._currentAdUnit.isShowing()) {
+            return;
+        }
+        this.shouldReinitialize().then((reinitialize) => {
+            if(reinitialize) {
+                if(this._currentAdUnit && this._currentAdUnit.isShowing()) {
+                    this._mustReinitialize = true;
+                    this.setRefreshAllowed(false);
+                } else {
+                    this._nativeBridge.Sdk.logDebug('Unity Ads webapp has been updated, reinitializing Unity Ads');
+                    this.reinitialize();
+                }
+            } else {
+                this.refresh();
+                this._sessionManager.sendUnsentSessions(this._operativeEventManager);
+            }
+        });
     }
 }